Adjust fuzzer warning and memory limits

+++ include/qpdf/QPDF.hh
@@ -228,9 +228,9 @@ class QPDF
     QPDF_DLL
     void setSuppressWarnings(bool);
 
-    // Set the maximum number of warnings to output. Subsequent warnings are suppressed.
+    // Set the maximum number of warnings. A QPDFExc is thrown if the limit is exceeded.
     QPDF_DLL
-    void setMaxWarnings(int);
+    void setMaxWarnings(size_t);
 
     // By default, QPDF will try to recover if it finds certain types of errors in PDF files.  If
     // turned off, it will throw an exception on the first such problem it finds without attempting
@@ -1501,7 +1501,7 @@ class QPDF
         bool provided_password_is_hex_key{false};
         bool ignore_xref_streams{false};
         bool suppress_warnings{false};
-        int max_warnings{0};
+        size_t max_warnings{0};
         bool attempt_recovery{true};
         bool check_mode{false};
         std::shared_ptr<EncryptionParameters> encp;
